requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
httpx==0.27.0
solana==0.34.0
solders==0.21.0
apscheduler==3.10.4
loguru==0.7.2
psutil==5.9.8
python-dotenv==1.0.1
rich==13.7.1
